refactor(Day3): move to single file with Container class
